Custom Button
feat: A new SwiftUI file created with the name CustomButtonView and added to Component folder.

feat: Styled the button and added a  system symbol to it.

feat: Replaced the text in the CardView to show/add the CustomButtonView to the card.

feat: Added a Spacer between the card title and the button.