Card Image
feat: Code added to generate a random new image on the card, every time that the "Explore More" button is pressed.

feat: An animation was added so that there is an ease-out animation when each new image loads.

feat: The code was modified for the GradientButton to swap the color gradient of the button when it's pressed.