Circle Customisation
feat: Function created to add random scale to the circles.

feat: Function created that generates a time interval. This is added as a modifier to the Circle. It adds a random speed to the circle animation.

feat: Function created that adds a random speed to the animated circles. A modifier added to the Circle that enables the speed.

feat: A new ZStack is added to the Circle in the CustomCircleView.